Collect Timedelta tests
