+++ pandas/tests/scalar/timedelta/test_constructors.py
@@ -8,11 +8,166 @@ from pandas._libs.tslibs import OutOfBoundsTimedelta
 from pandas._libs.tslibs.dtypes import NpyDatetimeUnit
 
 from pandas import (
+    Index,
     NaT,
     Timedelta,
+    TimedeltaIndex,
     offsets,
     to_timedelta,
 )
+import pandas._testing as tm
+
+
+class TestTimedeltaConstructorUnitKeyword:
+    @pytest.mark.parametrize("unit", ["Y", "y", "M"])
+    def test_unit_m_y_raises(self, unit):
+        msg = "Units 'M', 'Y', and 'y' are no longer supported"
+
+        with pytest.raises(ValueError, match=msg):
+            Timedelta(10, unit)
+
+        with pytest.raises(ValueError, match=msg):
+            to_timedelta(10, unit)
+
+        with pytest.raises(ValueError, match=msg):
+            to_timedelta([1, 2], unit)
+
+    @pytest.mark.parametrize(
+        "unit,unit_depr",
+        [
+            ("h", "H"),
+            ("min", "T"),
+            ("s", "S"),
+            ("ms", "L"),
+            ("ns", "N"),
+            ("us", "U"),
+        ],
+    )
+    def test_units_H_T_S_L_N_U_deprecated(self, unit, unit_depr):
+        # GH#52536
+        msg = f"'{unit_depr}' is deprecated and will be removed in a future version."
+
+        expected = Timedelta(1, unit=unit)
+        with tm.assert_produces_warning(FutureWarning, match=msg):
+            result = Timedelta(1, unit=unit_depr)
+        tm.assert_equal(result, expected)
+
+    @pytest.mark.parametrize(
+        "unit, np_unit",
+        [(value, "W") for value in ["W", "w"]]
+        + [(value, "D") for value in ["D", "d", "days", "day", "Days", "Day"]]
+        + [
+            (value, "m")
+            for value in [
+                "m",
+                "minute",
+                "min",
+                "minutes",
+                "Minute",
+                "Min",
+                "Minutes",
+            ]
+        ]
+        + [
+            (value, "s")
+            for value in [
+                "s",
+                "seconds",
+                "sec",
+                "second",
+                "Seconds",
+                "Sec",
+                "Second",
+            ]
+        ]
+        + [
+            (value, "ms")
+            for value in [
+                "ms",
+                "milliseconds",
+                "millisecond",
+                "milli",
+                "millis",
+                "MS",
+                "Milliseconds",
+                "Millisecond",
+                "Milli",
+                "Millis",
+            ]
+        ]
+        + [
+            (value, "us")
+            for value in [
+                "us",
+                "microseconds",
+                "microsecond",
+                "micro",
+                "micros",
+                "u",
+                "US",
+                "Microseconds",
+                "Microsecond",
+                "Micro",
+                "Micros",
+                "U",
+            ]
+        ]
+        + [
+            (value, "ns")
+            for value in [
+                "ns",
+                "nanoseconds",
+                "nanosecond",
+                "nano",
+                "nanos",
+                "n",
+                "NS",
+                "Nanoseconds",
+                "Nanosecond",
+                "Nano",
+                "Nanos",
+                "N",
+            ]
+        ],
+    )
+    @pytest.mark.parametrize("wrapper", [np.array, list, Index])
+    def test_unit_parser(self, unit, np_unit, wrapper):
+        # validate all units, GH 6855, GH 21762
+        # array-likes
+        expected = TimedeltaIndex(
+            [np.timedelta64(i, np_unit) for i in np.arange(5).tolist()],
+            dtype="m8[ns]",
+        )
+        # TODO(2.0): the desired output dtype may have non-nano resolution
+        msg = f"'{unit}' is deprecated and will be removed in a future version."
+
+        if (unit, np_unit) in (("u", "us"), ("U", "us"), ("n", "ns"), ("N", "ns")):
+            warn = FutureWarning
+        else:
+            warn = None
+        with tm.assert_produces_warning(warn, match=msg):
+            result = to_timedelta(wrapper(range(5)), unit=unit)
+            tm.assert_index_equal(result, expected)
+            result = TimedeltaIndex(wrapper(range(5)), unit=unit)
+            tm.assert_index_equal(result, expected)
+
+            str_repr = [f"{x}{unit}" for x in np.arange(5)]
+            result = to_timedelta(wrapper(str_repr))
+            tm.assert_index_equal(result, expected)
+            result = to_timedelta(wrapper(str_repr))
+            tm.assert_index_equal(result, expected)
+
+            # scalar
+            expected = Timedelta(np.timedelta64(2, np_unit).astype("timedelta64[ns]"))
+            result = to_timedelta(2, unit=unit)
+            assert result == expected
+            result = Timedelta(2, unit=unit)
+            assert result == expected
+
+            result = to_timedelta(f"2{unit}")
+            assert result == expected
+            result = Timedelta(f"2{unit}")
+            assert result == expected
 
 
 def test_construct_from_kwargs_overflow():
